requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.7
django-extensions==3.2.1
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.5
pytest==7.2.1
pytest-django==4.5.2
pytz==2022.7.1
sqlparse==0.4.3
